feat: extracts hofhoc for routes with tabs
